refactor(sharing-dialog): move components into subfolders
