+++ components/sharing-dialog/src/access-select/custom-single-select-option.js
