+++ components/sharing-dialog/src/access-select/index.js
@@ -1,0 +1,1 @@
+export { AccessSelect } from './access-select.js'
