+++ components/sharing-dialog/src/allow-access/__tests__/allow-access.test.js
@@ -2,13 +2,13 @@ import { Button } from '@dhis2-ui/button'
 import { CustomDataProvider } from '@dhis2/app-runtime'
 import { mount } from 'enzyme'
 import React from 'react'
-import { AccessSelect } from '../access-select.js'
-import { Autocomplete } from '../autocomplete/autocomplete.js'
-import { ShareBlock } from '../share-block.js'
+import { AccessSelect } from '../../access-select/index.js'
+import { Autocomplete } from '../../autocomplete/index.js'
+import { AllowAccess } from '../allow-access.js'
 
-describe('<ShareBlock />', () => {
+describe('<AllowAccess />', () => {
     it('renders a single Autocomplete component ', () => {
-        const wrapper = mount(<ShareBlock onAdd={() => {}} />, {
+        const wrapper = mount(<AllowAccess onAdd={() => {}} />, {
             wrappingComponent: CustomDataProvider,
         })
         const autocompletes = wrapper.find(Autocomplete)
@@ -17,7 +17,7 @@ describe('<ShareBlock />', () => {
     })
 
     it('renders a single AccessSelect component', () => {
-        const wrapper = mount(<ShareBlock onAdd={() => {}} />, {
+        const wrapper = mount(<AllowAccess onAdd={() => {}} />, {
             wrappingComponent: CustomDataProvider,
         })
         const accessSelects = wrapper.find(AccessSelect)
@@ -26,7 +26,7 @@ describe('<ShareBlock />', () => {
     })
 
     it('renders a button with "Give access"', () => {
-        const wrapper = mount(<ShareBlock onAdd={() => {}} />, {
+        const wrapper = mount(<AllowAccess onAdd={() => {}} />, {
             wrappingComponent: CustomDataProvider,
         })
         const buttons = wrapper.find(Button)
