+++ components/sharing-dialog/src/allow-access/allow-access.js
@@ -3,11 +3,11 @@ import { useDataQuery, useOnlineStatus } from '@dhis2/app-runtime'
 import { colors } from '@dhis2/ui-constants'
 import PropTypes from 'prop-types'
 import React, { useMemo, useState } from 'react'
-import { AccessSelect } from './access-select.js'
-import { Autocomplete } from './autocomplete/autocomplete.js'
-import { debounce } from './helpers.js'
-import i18n from './locales/index.js'
-import { ACCESS_VIEW_ONLY, ACCESS_VIEW_AND_EDIT } from './sharing-constants.js'
+import { AccessSelect } from '../access-select/index.js'
+import { Autocomplete } from '../autocomplete/index.js'
+import { debounce } from '../helpers.js'
+import i18n from '../locales/index.js'
+import { ACCESS_VIEW_ONLY, ACCESS_VIEW_AND_EDIT } from '../sharing-constants.js'
 
 const query = {
     usersAndGroups: {
@@ -20,7 +20,7 @@ const query = {
 
 const addType = (type) => (result) => ({ ...result, type })
 
-export const ShareBlock = ({ onAdd }) => {
+export const AllowAccess = ({ onAdd }) => {
     const [userOrGroup, setUserOrGroup] = useState(undefined)
     const [access, setAccess] = useState(undefined)
     const [didInvalidate, setDidInvalidate] = useState(true)
@@ -150,6 +150,6 @@ export const ShareBlock = ({ onAdd }) => {
     )
 }
 
-ShareBlock.propTypes = {
+AllowAccess.propTypes = {
     onAdd: PropTypes.func.isRequired,
 }
