+++ components/sharing-dialog/src/allow-access/index.js
@@ -1,0 +1,1 @@
+export { AllowAccess } from './allow-access.js'
