+++ components/sharing-dialog/src/autocomplete/index.js
@@ -1,0 +1,1 @@
+export { Autocomplete } from './autocomplete.js'
